requirements.txt
# Core
kubernetes>=28.0.0
pyyaml>=6.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0

# Monitoring
prometheus-client>=0.19.0
prometheus-api-client>=0.5.3

# Web
flask>=3.0.0

# Load testing
locust>=2.20.0

# Logging
loguru>=0.7.0

# HTTP
httpx>=0.25.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0